move regionmaskcompat to core
