+++ mesmer/core/regionmaskcompat.py
